Cmplying to third review

+++ benchmark/bench-fft.cpp
@@ -327,7 +327,7 @@ static void BM_InvFFTNativeRadix2CopyScaled(
   const size_t fft_size = state.range(0);
   const size_t bound = 1 << 30;
   const double scale = 1.3611294676837539e+39;  // (1 << 130)
-  const double inv_scale = static_cast<double>(1.0) / scale;
+  const double inv_scale = 1.0 / scale;
   FFT fft(fft_size, nullptr);
 
   AlignedVector64<std::complex<double>> output(fft_size);
@@ -500,7 +500,7 @@ static void BM_InvFFTAVX512InPlaceScaled(benchmark::State& state) {  //  NOLINT
   const size_t fft_size = state.range(0);
   const size_t bound = 1 << 30;
   const double scale = 1.3611294676837539e+39;  // (1 << 130)
-  const double inv_scale = static_cast<double>(1.0) / scale;
+  const double inv_scale = 1.0 / scale;
   FFT fft(fft_size, nullptr);
 
   AlignedVector64<double> input(2 * fft_size);
